Employment (by Place of Work, by Place of Residence) by Industry data

Total employment by industry, by place of work and by place of usual residence (based on Local Government Area boundies) for ABS Census years 2011, 2016 and 2021. From ABS TableBuilder Pro. Only includes employment where people live and work in the same LGA.

Usage

live_and_work

Format

live_and_work

A data frame with 36,564 rows and 5 columns.

industry

ANZSIC 19 Sector Industry

lga_ur

Place (Local Government Area) of usual residency

lga_pow

Place (Local Government Area) of work

employment

Total employment

year

Census year