Generate Data For Regional Input-Output Table Creation

get_data creates a list of data frames containing all the data necessary to build a Regional Input-Output Table.

Usage

get_data(year)

Arguments

year

Census year

Value

Details

Regional Input-Output Tables are created using the Location Quotient method and are derived from the Australian 19 Sector Input-Output Table. This function provides:

Australian 19 Sector Input-Output Table

Industry Productivity

FTE Ratios

Regional Location Quotients

Australian 19 Sector Input-Output Table

The 19 Sector Input-Output Table for Australia is derived from the Australian National Accounts: Input-Output Tables. The Input-Output Table provided by the Australian Bureau of Statistics is more detailed - describing industry-industry flows for 114 sectors. As Regional Input-Output Tables are estimated using the Location Quotient method, an aggregated 19 Sector Input-Output Table is used which is a trade-off between availability and reliability of regional data, and specificity of the regional model.

Industry Productivity

Industry productivity measures the average production per FTE employment across Australia.

FTE Ratios

The ratio of FTE Employment to Total Employment across all industries in Australia.

Regional Location Quotients

Location Quotients are used to determine if an industry in a region is "as significant" as the industry in the country. That is, whether or not a region capable of supplying to local industries at the same proportion as the country as a whole. If not, the region supplies proportionally less to local industries. Regional employment is used to determine whether or not an industry in a region is "significant" where significance is defined as:

\[ LQ_{i,r} = \frac{E_{i, r}}{\sum_i{E_{i, r}}} / \frac{\sum_{r}{E_{i, r}}}{\sum_{r,i}{E_{i,r}}} \] Where \(E_{i,r}\) is the FTE Employment in region r, and industry i.

An industry in a region is said to be significant if \(LQ_{i,r} >=1\). In these instances, the regional coefficient for a supplying industry is assumed to be the same as the national coefficient for the supplying industry. Otherwise, the regional coefficient is estimated as \(LQ_{i,r}\times a_{i,r}\) where \(a_{i,r}\) is the national coefficient.

Examples

get_data(2021)